refactor: move actor top picks use cases to toppicks package
