+++ presentation/src/main/java/com/london/presentation/feature/details/actor/ActorDetailsViewModel.kt
@@ -3,8 +3,8 @@ package com.london.presentation.feature.details.actor
 import androidx.lifecycle.SavedStateHandle
 import com.london.domain.usecase.GetActorDetailsByIdUseCase
 import com.london.domain.usecase.GetActorImagesByIdUseCase
-import com.london.domain.usecase.GetActorMoviePicksByIdUseCase
-import com.london.domain.usecase.GetActorTvShowPicksByIdUseCase
+import com.london.domain.usecase.toppicks.GetActorMoviePicksByIdUseCase
+import com.london.domain.usecase.toppicks.GetActorTvShowPicksByIdUseCase
 import com.london.presentation.feature.base.BaseViewModel
 import com.london.presentation.navigation.Screen
 import com.london.presentation.navigation.getArgs
